AP_Radio: switch to 16 bit TX version
using year, month, day

+++ libraries/AP_Radio/AP_Radio_backend.h
@@ -55,7 +55,7 @@ public:
     virtual void update(void) = 0;
 
     // get TX fw version
-    virtual uint8_t get_tx_version(void) = 0;
+    virtual uint32_t get_tx_version(void) = 0;
     
     // get radio statistics structure
     virtual const AP_Radio::stats &get_stats(void) = 0;
